refactor(voice): split factory.rs into factory/ directory
Split the 1279-line factory.rs into five focused sub-modules
(traits, stt_providers, tts_providers, helpers, entry) plus a
companion tests.rs, each under 500 lines. Public API preserved
exactly via pub use re-exports in mod.rs.
